ecommerce-rdb/requirements.txt
Faker==37.1.0
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.3
PyJWT==2.8.0
sqlmodel==0.0.24
uvicorn==0.34.1
